# Allow `%Currency{}` structs in `currency_for_code/3`