This PR is part of a series of PRs that aim to improve performance by allowing both `%Cldr.Number.Format.Options{}` structs to be passed as options in place of keyword lists, and to add support for `%Cldr.Currency{}` structs as alternative for currency codes.

This PR adds support to pass a `%Cldr.Currency` struct directly, and will return it if it's passed to `currency_for_code/3`. Since `currency_for_code/3` is relatively complex, and iterates over currencies (which may be something that can be addressed in a future PR, replacing it with a keyed lookup via a map), this change attempts to offer developers a way to pass the currency struct, already resolved. This in turn made a real-world impact when we ran this in a tight loop. In our use-case, we know the currencies beforehand, and we pre-resolve these currencies. We take the trade-off between slighly higher compile time, and much faster runtime performance.


I ran this benchmark:
https://gist.github.com/jeroenvisser101/32deb85f50fe76f8cd5d48049591e355

```
Operating System: macOS
CPU Information: Intel(R) Xeon(R) W-3245 CPU @ 3.20GHz
Number of Available Cores: 32
Available memory: 160 GB
Elixir 1.12.0
Erlang 24.0

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 2 s
parallel: 1
inputs: none specified
Estimated total run time: 28 s

Benchmarking after...
Benchmarking before...

Name             ips        average  deviation         median         99th %
after       104.36 K        9.58 μs   ±273.14%           8 μs          22 μs
before        2.36 K      423.16 μs    ±33.75%         410 μs         576 μs

Comparison: 
after       104.36 K
before        2.36 K - 44.16x slower +413.58 μs

Memory usage statistics:

Name      Memory usage
after          9.51 KB
before       217.17 KB - 22.84x memory usage +207.66 KB

**All measurements for memory usage were the same**
```