@kipcole9 2021-06-16
Thanks for the PR, much appreciated. Great suggestion.  Its the same principle that led me to allow precomputed options to be passed to `Cldr.Number.to_string/2` since in tight loops, options processing is 50% of the compute time. And clearly from your benchmark its worse than that for currencies.

I will definitely look into whether currency validation can be improved as you suggest but this PR is a great start.