@jeroenvisser101 2021-06-16
Thanks! We use Money and Cldr very heavily, and appreciate the work you put into this, it has never let us down. If I can find the time, I might be able to contribute more performance related enhancements :)