@kipcole9 2021-06-17
Published [ex_cldr_currencies version 2.10.0](https://hex.pm/packages/ex_cldr_currencies/2.10.0) with the following changelog:

#### Enhancements

* Allow `%Currency{}` structs in `currency_for_code/3`. Thanks to @jeroenvisser101 for the PR. [Improves the performance](https://github.com/elixir-cldr/cldr_currencies/pull/4) by up to 40x when validating currencies when the currency has already been pre-constructed. 

* Add implementation of the `Inspect` protocol for `t:Cldr.Currency` structs.

If there are other particular hotspots you find, please keep the issues coming!